Refactor and remove EventPrototype class
