--- Assets/Editor/EventPrototypeEditorTest.cs
@@ -1,19 +1,0 @@
-﻿#region License
-// ====================================================
-// Project Porcupine Copyright(C) 2016 Team Porcupine
-// This program comes with ABSOLUTELY NO WARRANTY; This is free software, 
-// and you are welcome to redistribute it under certain conditions; See 
-// file LICENSE, which is part of this source code package, for details.
-// ====================================================
-#endregion
-using System;
-using NUnit.Framework;
-using Scheduler;
-
-public class EventPrototypeEditorTest
-{
-    [Test]
-    public void EditorTest()
-    {
-    }
-}
