--- Assets/Scripts/Scheduler/EventPrototype.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 259c3ea9b3156f141a668f83319d9415
-timeCreated: 1472427395
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
